telegram_kimi_bot/requirements.txt
python-telegram-bot>=20.7
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
